requirements.txt
streamlit==1.32.0
fastapi==0.110.0
uvicorn==0.27.1
pandas==2.2.0
python-dotenv==1.0.0
langchain==0.3.23
langchain-huggingface==0.0.6
huggingface_hub==0.21.3
sentence-transformers==2.3.1
faiss-cpu==1.7.4
